[pass] Move postgres-dependent tests to e2e/postgres directory.
Problems:
- postgres tests in unit/ fail without live database in CI
- unit tests should mock DBs per project conventions

Solutions:
- moved 3 test files to cognee/tests/e2e/postgres/
- created __init__.py for e2e/ and e2e/postgres/
- updated import path in test_graphdb_postgres.py
- updated 2 test paths in graph_db_tests.yml

Architectural Trade-Off Decisions:
- chose e2e/ over integration/ to avoid automatic CI sweep

Reasoning and Related Notes:
- requested by Igor (dexters1) during PR #2506 review
- mirrors identical move on sister hybrid-adapter branch
